Renamed error file to errorReport
